Add RepoGetter.py to retrieve GitHub clone links
This commit adds a new file, RepoGetter.py, to the ocr directory. The file contains a function, get_github_clone_links, that retrieves the clone links of a user's GitHub repositories using the GitHub API. It also includes a script that demonstrates the usage of the function by printing the clone links of a user specified by the 'username' variable.